refactor(lsp): split inlay hint link suggester
